pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "agno>=1.2.15",
    "aiofiles>=24.1.0",
    "arxiv>=2.2.0",
    "baidusearch>=1.0.3",
    "bcrypt>=4.3.0",
    "black>=25.1.0",
    "duckduckgo-search>=8.0.0",
    "email-validator>=2.2.0",
    "fastapi[all]>=0.115.12",
    "google-adk>=0.5.0",
    "google-api-python-client>=2.169.0",
    "google-genai>=1.10.0",
    "googlesearch-python>=1.3.0",
    "matplotlib>=3.10.3",
    "numpy>=2.2.4",
    "passlib>=1.7.4",
    "pillow>=11.1.0",
    "pip-audit>=2.9.0",
    "pycountry>=24.6.1",
    "pydantic-settings>=2.8.1",
    "pydantic[email]>=2.11.2",
    "pypdf>=5.4.0",
    "python-dotenv>=1.1.0",
    "python-jose[cryptography]>=3.4.0",
    "python-multipart>=0.0.20",
    "scikit-image>=0.25.2",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.0",
]
